social_api/serializers.py
from django.contrib.auth import get_user_model
from rest_framework import serializers
from .models import FriendRequest

CustomUser = get_user_model()


class FriendRequestSerializer(serializers.ModelSerializer):
    class Meta:
        model = FriendRequest
        fields = '__all__'
        read_only_fields = ['sender', 'receiver', 'status']


class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = CustomUser
        fields = ['id', 'username', 'email', 'first_name', 'last_name']
